make board demo can be accessed by non-privileged client. update url format to be restful
